Use technically correct Haskell bind names
This possibly is nitpicking, but there's no `bind` method in Haskell's `Monad` type class. Technically, its name is `>>=`. I think that most Haskellers call it _bind_ in their heads, so there's nothing in the present text that would confuse a Haskell programmer.

On the other hand, for the benefit of a reader who comes to this document from F# and wants to look up the corresponding Haskell functions, it could perhaps be confusing if he or she goes looking for a `bind` function that doesn't exist.