Clarify relationship to maths notion of monoid
After reading this document, I wanted to play around with the monoidal composition features of computation expressions. I soon ran into trouble implementing `Yield` for an endomorphism, and that made me reread the document. I realised that I'd misunderstood the section on monoidal composition.

Perhaps these edits make the relationship with mathematical monoids clearer. I first misread the section in such a way that I thought I could get syntactic sugar over _any_ monoid, but that's not the case. The syntax is 'only' available for monoids on monads.